requirements.txt
chromadb==1.5.5
sentence-transformers==5.3.0
pdfplumber==0.11.9
fastapi==0.135.1
uvicorn==0.42.0
requests==2.32.5
python-dotenv==1.2.2
